use the revised QMutexPool API
[git-p4: depot-paths = "//depot/qt/main/": change = 65953]

+++ src/tools/qlibrary_mac.cpp
@@ -108,7 +108,7 @@ bool QLibraryPrivate::loadLibrary()
 
 #ifdef QT_THREAD_SUPPORT
     // protect glibs_loaded creation/access
-    QMutexLocker locker( qt_global_mutexpool->get( (int)&glibs_loaded ) );
+    QMutexLocker locker( qt_global_mutexpool->get( &glibs_loaded ) );
 #endif // QT_THREAD_SUPPORT
 
 #if defined(QT_DEBUG) || defined(QT_DEBUG_COMPONENT)
@@ -160,7 +160,7 @@ bool QLibraryPrivate::freeLibrary()
 
 #ifdef QT_THREAD_SUPPORT
     // protect glibs_loaded access
-    QMutexLocker locker( qt_global_mutexpool->get( (int)&glibs_loaded ) );
+    QMutexLocker locker( qt_global_mutexpool->get( &glibs_loaded ) );
 #endif // QT_THREAD_SUPPORT
 
     if(glibs_loaded) {
